Vario acoustic feedback: proof-of-concept
This is an attempt to implement acoustic feedback for clibm rate.
Enable this by adding a 16-bit sensor of type IBUS_MEAS_TYPE_CLIMB_RATE
(= 0x09), and setting one of the alarms (no matter what limits)
for this sensor in the Extra menu.

I wanted the feedback to be quiet in order not to override other alarms.
So I make very short beep at 1024 Hz followed by longer beep of
higher frequency when climbing (up to 2048 Hz) or lower frequency
when descending (down to 512 Hz).

This is a proof of concept only, for review. The better solution
would probably be

- calculate the rate directly from the altitude or pressure sensor
	just before beeping, in order to lower the latency
	(currently it lags about 1 second on average after the real
	climb or descend takes place).

- add an Extra menu item separately for vario feedback, allowing to
	select any sensor as a basis for the feedback. I think only
	GPS Altitude, Altitude, and Pressure make sense, but maybe
	someone can find it useful to set up vario feedback for
	RPM, Current, or Battery voltage.

- evaluate the vario feedback after all alarms, and only if no alarm
	has been triggered.